PriceList/requirements.txt
kivy==2.2.1
pdfminer.six==20221105